fixed kernel mode. working

+++ src/servo/servo.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python3
 from .PWM import PWM
 import time
+import os.path
 
 
 class Servo:
@@ -8,9 +9,9 @@ class Servo:
     channel = 0
 
     def __init__(self, channel):
+        self.channel = channel
         if not self.init:
             self.init_controller()
-        self.channel = channel
 
     def init_controller(self):
         raise NotImplementedError
@@ -22,8 +23,8 @@ class Servo:
 class ServoPython(Servo):
     pwm = PWM(0x40)
     init = False
-    PULSE_MIN = 150  # Min pulse length out of 4096
-    PULSE_MAX = 500  # Max pulse length out of 4096
+    PULSE_MIN = 150
+    PULSE_MAX = 500
 
     def init_controller(self):
         self.pwm.set_pwm_freq(60)
@@ -39,24 +40,31 @@ class ServoPython(Servo):
 
 
 class ServoKernel(Servo):
-    period_ns = 20000000
+    period_ns = 20000000.0
     servoClassPath = '/sys/class/pwm/pwmchip0'
-    model_slope = 9268.3
-    model_intercept = 1550000 
-    PULSE_MIN = 600000
-    PULSE_MAX = 2500000
+    model_slope = 531034.0
+    model_intercept = 1550000.0 
+    PULSE_MIN = 600000.0
+    PULSE_MAX = 2500000.0
     duty_cycle_path = ''
 
     def init_controller(self):
         export_path = "{:s}/export".format(self.servoClassPath)
-        with open(export_path, 'w') as fh:
-            fh.writelines("{:d}".format(self.channel))
-
         pwm_path = "{:s}/pwm{:d}".format(self.servoClassPath, self.channel)
+        if not os.path.exists(pwm_path):
+            try:
+                with open(export_path, 'w') as fh:
+                    fh.writelines("{:d}".format(self.channel))
+            except OSError as err:
+                print("Error in writing to " + export_path)
+                print(self.channel)
+                print(err)
+                exit(0)
+
         period_path = "{:s}/period".format(pwm_path)
         self.duty_cycle_path = "{:s}/duty_cycle".format(pwm_path)
         with open(period_path, 'w') as fh:
-            fh.writelines("{:d}".format(self.period_ns))
+            fh.writelines("{:d}".format(int(self.period_ns)))
 
     def move_to_angle(self, radian):
         duty_cycle = self.model_slope*radian + self.model_intercept
@@ -65,12 +73,13 @@ class ServoKernel(Servo):
         elif duty_cycle > self.PULSE_MAX:
             duty_cycle = self.PULSE_MAX
         with open(self.duty_cycle_path, 'w') as fh:
-            fh.writelines("{:d}".format(duty_cycle))
+            #print(duty_cycle)
+            fh.writelines("{:d}".format(int(duty_cycle)))
 
 
 if __name__ == '__main__':
-    s0 = Servo(0)
-    s1 = Servo(2)
+    s0 = ServoKernel(0)
+    s1 = ServoKernel(2)
     wt = 0.20
     while True:
         s0.move_to_angle(0)
